requirements.txt
# Core
python-dotenv
mcp
fastapi
uvicorn
pydantic
fastmcp

# LLM
google-generativeai
anthropic

# Vision
mss
pillow

# MCP Server Libraries (for dynamic tool creation)
pyautogui
pydirectinput
pygetwindow
pyperclip
psutil
easyocr
opencv-python
numpy
pywin32
